+++ src/aioverse/__init__.py
@@ -1,19 +1,13 @@
 # 服务模块
 from . import OpenAI
-from . import Log
 
 # 基础模块 协议接口
 from . import errors
 from . import protocols
-from . import managers
-from . import utils
 
 
 __all__ = [
 	"OpenAI",
-	"Log",
-	"managers",
 	"errors",
 	"protocols",
-	"utils"
 ]