+++ src/aioverse/models/contexts/user.py
@@ -1,8 +1,0 @@
-from .base_context	import Context
-
-from typing	import Literal
-
-
-class User(Context):
-	
-	role: Literal["user"] = "user"