+++ src/aioverse/models/response/usage.py
@@ -1,5 +1,5 @@
 from pydantic	import BaseModel, Field
-from typing		import Dict
+from typing		import Dict, Optional
 
 
 class Usage(BaseModel):
@@ -7,5 +7,5 @@ class Usage(BaseModel):
 	completion_tokens			: int
 	prompt_tokens				: int
 	total_tokens				: int
-	completion_tokens_details	: Dict[str, int] = Field(default_factory=dict)
-	prompt_tokens_details		: Dict[str, int] = Field(default_factory=dict)
+	completion_tokens_details	: Optional[Dict[str, int]] = Field(default_factory=dict)
+	prompt_tokens_details		: Optional[Dict[str, int]] = Field(default_factory=dict)