+++ src/aioverse/models/segments/__init__.py
@@ -1,16 +1,22 @@
-from .base_segment		import Segment
-from .audio_segment		import AudioInput
-from .image_url_segment	import ImageUrl
-from .text_segment		import Text
+from .base_segment			import BaseSegment
+from .audio_segment			import AudioInputSegment
+from .file_segment			import FileSegment
+from .image_base64_segment	import ImageBase64Segment
+from .image_url_segment		import ImageUrlSegment
+from .text_segment			import TextSegment
+from .unknown_segment		import UnknownSegment
 
 
 __all__ = [
 	
 	# base
-	"Segment",
+	"BaseSegment",
 	
-	"AudioInput",
-	"ImageUrl",
-	"Text"
+	"AudioInputSegment",
+	"FileSegment",
+	"ImageBase64Segment",
+	"ImageUrlSegment",
+	"TextSegment",
+	"UnknownSegment"
 	
 ]