+++ src/aioverse/models/tool_calling/tool_calling.py
@@ -1,10 +1,12 @@
-from pydantic	import BaseModel
+from pydantic	import BaseModel, Field
+from typing		import Optional
 
 from .function	import Function
 
 
 class ToolCalling(BaseModel):
 
-	id		: str
-	type	: str
-	function: Function
+	index	: Optional[int]	= Field(default=None)
+	id		: str			= Field(default="")
+	type	: str			= Field(default="function")
+	function: Function		= Field(default_factory=Function)