+++ src/aioverse/models/tool_schema/tool.py
@@ -1,9 +1,0 @@
-from pydantic	import BaseModel
-
-from .function	import Function
-
-
-class Tool(BaseModel):
-	
-	type	: str = "function"
-	function: Function