+++ src/aioverse/utils/__init__.py
@@ -1,9 +1,0 @@
-# 语法糖
-from . import syntax_sugar
-# 占位
-from . import holder
-
-
-__all__ = [
-	"syntax_sugar"
-]